Undo polyglotism
I don't plan to be writing a whole lot of these parsers.
This repo's just gonna be for my Java fractional calculator.
I removed the `wtf_ja` instructions because nobody in their left
mind would ever use that. Added a makefile and fixed a small bug
with the color of the text not being reset at exit.
Still super proud of this.